this is an add-on to the previous commit message so hence the white space changes
Fixed the tests for the arduino controller. This took a lot more time than expected and required multiple people helping in order to get the fix working. Initially there was an error about a query returning multiple values when one was expected. it took over an hour to finally find that this was referring to multiple gardens being returned in the findAllByArduinoId function. After fixing this there was another problem with multiple gardens being created in the test class which meant that when each test ran it was searching for data linked to the newest garden but the data was linked to the garden made in the first test. This was fixed by changing the if statement to check if gardens had already been created and using those if they existed instead of creating a new one. This all took a lot of debugging and a lot of time was spent scratching our heads and going over the same test many times before finding something to investigate further. Along with these testing changes I also spent time to figure out why when running tests as "intergration", "unit" and "cucumber" still ran all the tests on each. Turns out, in the build.grade file there is a the function useJUnitPlatform(), which, under the hood, runs all tests by default. This means we have to exclude and include the right tests within the filter stating includeTestsMatching ... and excludeTestsMatching ... for each test required